fix: int64/int mixup, missing imports

+++ moonlight/value_golua.go
@@ -24,8 +24,8 @@ func StringValue(str string) Value {
 	return rt.StringValue(str)
 }
 
-func IntValue(i int) Value {
-	return rt.IntValue(int64(i))
+func IntValue(i int64) Value {
+	return rt.IntValue(i)
 }
 
 func BoolValue(b bool) Value {
